test_requirements.txt
astroid==1.5.3
attrs==17.3.0
certifi==2017.11.5
chardet==3.0.4
flake8==3.5.0
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.7.5
pytest==3.3.0
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
zope.interface==4.4.3
